Avoid overscroll on editor page
Prior to this commit, if you scrolled on the navigation bar at the top
of the editor screen, it'd do that little overscrolling thing which
browsers do when you try to scroll past the end of a page. This felt
super weird because we normally don't scroll this page at all.